# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Discord Bot Hosting Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to the panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /env-vars:
    get:
      operationId: listEnvVars
      tags: [env-vars]
      summary: List all environment variables
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Environment variables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnvVar"
    post:
      operationId: createEnvVar
      tags: [env-vars]
      summary: Create an environment variable
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVar"
  /env-vars/{id}:
    put:
      operationId: updateEnvVar
      tags: [env-vars]
      summary: Update an environment variable
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVar"
    delete:
      operationId: deleteEnvVar
      tags: [env-vars]
      summary: Delete an environment variable
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get panel configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanelConfig"
    put:
      operationId: updateConfig
      tags: [config]
      summary: Update panel configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdate"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanelConfig"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthToken:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [access_token, token_type, user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, username, is_admin, created_at]
...
    EnvVar:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        is_secret:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, key, value, is_secret, created_at, updated_at]
    EnvVarInput:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        is_secret:
          type: boolean
      required: [key, value]
    EnvVarUpdate:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        is_secret:
          type: boolean
...
    PanelConfig:
      type: object
      properties:
        bot_name:
          type: string
        main_file:
          type: string
        auto_restart:
          type: boolean
        auto_install_deps:
          type: boolean
        max_log_lines:
          type: integer
        python_version:
          type: string
      required: [bot_name, main_file, auto_restart, auto_install_deps, max_log_lines, python_version]
    ConfigUpdate:
      type: object
      properties:
        bot_name:
          type: string
        main_file:
          type: string
        auto_restart:
          type: boolean
        auto_install_deps:
          type: boolean
        max_log_lines:
          type: integer
        python_version:
          type: string
...